Network Admin: sites list
In Jetpack version over 4.0, the remote registration action requires
an additional 'state' parameter so that secrets can be tied to users.

This was handled in `Jetpack::register`, but was missed in the network admin.

This PR refactors `Jetpack::generate_secrets` so that it is used consistently
between single site dashboards, and the network admin.